+++ premake5.lua
@@ -216,6 +216,7 @@ group "exts"
     include ("source/extensions/omni.usd.schema.isaac")
     include ("source/extensions/omni.isaac.dr")
     include ("source/extensions/omni.isaac.dynamic_control")
+    include ("source/extensions/omni.isaac.dynamic_control.samples")
     include ("source/extensions/omni.isaac.contact_sensor")
     include ("source/extensions/omni.isaac.range_sensor")
     include ("source/extensions/omni.isaac.manip")