+++ source/extensions/omni.isaac.dynamic_control/python/tests/test_core.py
@@ -12,7 +12,7 @@
 #   For most things refer to unittest docs: https://docs.python.org/3/library/unittest.html
 import omni.kit.test
 
-import omni.kit.usd
+import omni.usd
 
 # Import extension python module we are testing with absolute import path, as if we are external user (other extension)
 from omni.isaac.dynamic_control import _dynamic_control